cache generated integrationtest domain classes
invalidate cache based on the codegen source, test schemas and (this) build.sbt

+++ integration-tests/schemas/build.sbt
@@ -4,11 +4,31 @@ val generateDomainClasses = taskKey[Seq[File]]("generate overflowdb domain class
 
 // cross scalaVersion is defined in project/Build.scala
 
-generateDomainClasses := Def.task {
+/** generated results are cached, based on the codegen source, test schemas and (this) build.sbt */
+generateDomainClasses := Def.taskDyn {
   val outputRoot = target.value / "odb-codegen"
-  FileUtils.deleteRecursively(outputRoot)
-  (Compile/runMain).toTask(s" CodegenForAllSchemas").value
-  FileUtils.listFilesRecursively(outputRoot)
+  val currentSchemaMd5 = FileUtils.md5(
+    sourceDirectory.value,
+    file("codegen/src/main"),
+    file("integration-tests/schemas/build.sbt"))
+
+  if (outputRoot.exists && lastSchemaMd5 == Some(currentSchemaMd5)) {
+    Def.task {
+      // inputs did not change, don't regenerate
+      FileUtils.listFilesRecursively(outputRoot)
+    }
+  } else {
+    Def.task {
+      FileUtils.deleteRecursively(outputRoot)
+      val invoked = (Compile/runMain).toTask(s" CodegenForAllSchemas").value
+      lastSchemaMd5(currentSchemaMd5)
+      FileUtils.listFilesRecursively(outputRoot)
+    }
+  }
 }.value
 
 publish/skip := true
+
+lazy val schemaMd5File                 = file("target/schema-src.md5")
+def lastSchemaMd5: Option[String]      = scala.util.Try(IO.read(schemaMd5File)).toOption
+def lastSchemaMd5(value: String): Unit = IO.write(schemaMd5File, value)
